refactor: change structure of the phrase-wiz game
