+++ src/components/pages/phrase-wizard-page/app/button.js
@@ -1,8 +1,8 @@
 import { createElement } from 'helpers/dom';
-import GameWords from 'components/phrase-wizard-page/app/words';
-import Service from 'components/phrase-wizard-page/app/service';
-import Results from 'components/phrase-wizard-page/app/results';
-import Voice from 'components/phrase-wizard-page/app/voice';
+import GameWords from 'components/pages/phrase-wizard-page/app/words';
+import Service from 'components/pages/phrase-wizard-page/app/service';
+import Results from 'components/pages/phrase-wizard-page/app/results';
+import Voice from 'components/pages/phrase-wizard-page/app/voice';
 
 export default class Button {
   static render() {
