+++ src/components/pages/phrase-wizard-page/app/service.js
@@ -1,6 +1,6 @@
 import { User, Words } from 'services/backend';
 import Toaster from 'components/Toaster';
-import GameWords from 'components/phrase-wizard-page/app/words';
+import GameWords from 'components/pages/phrase-wizard-page/app/words';
 
 export default class Service {
   static async wordsRequest(level) {
