+++ src/components/pages/phrase-wizard-page/app/statisctic.js
